requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
black==25.1.0
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography>=41.0.1
--no-cache-dir
distlib==0.3.9
distro==1.9.0
Django==5.1.6
dnspython==2.6.1
ecdsa==0.19.1
fastapi==0.115.14
filelock==3.13.1
fire==0.7.0
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.7
frozenlist==1.4.1
fsspec==2025.5.1
future==1.0.0
gast==0.5.4
gevent==24.2.1
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.16.0
h5py==3.10.0
httpcore==1.0.9
httpx==0.27.0
huggingface-hub==0.33.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==2.15.0
libclang==18.1.1
llama_stack==0.2.12
llama_stack_client==0.2.12
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.0.5
mypy_extensions==1.1.0
namex==0.0.7
numpy==1.26.4
oauthlib==3.2.2
openai==1.93.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt-einsum==3.3.0
optree==0.10.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.3.0
pathspec==0.12.1
pillow==10.2.0
pipenv==2024.4.1
platformdirs==4.3.6
prompt_toolkit==3.0.51
protobuf==4.25.3
psycopg2-binary==2.9.10
pyaml==25.5.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==4.6.3
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.4.0
rich==13.7.1
rpds-py==0.26.0
rsa==4.9
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.3
starlette==0.46.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tiktoken==0.9.0
tldextract==5.1.2
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.14.0
tzdata==2025.1
urllib3==2.2.1
virtualenv==20.29.2
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.2
